item request

+++ 05 - Bills/Bills/Bills.Services/ItemService.cs
@@ -1,9 +1,10 @@
-﻿namespace Bills.Services;
+﻿using Bills.Core.DTO.Requests;
+
+namespace Bills.Services;
 public class ItemService(AppDbContext dbContext) : IItemService
 {
-    public async Task<ErrorOr<ItemModel>> CreateAsync(ItemModel item)
+    public async Task<ErrorOr<ItemModel>> CreateAsync(ItemModelRequest item)
     {
-        //nincs ellenorzes h letezik-e
         var newItem = item.ToEntity();
 
         await dbContext.Items.AddAsync(newItem);
@@ -14,6 +15,13 @@ public class ItemService(AppDbContext dbContext) : IItemService
 
     public async Task<ErrorOr<Success>> DeleteAsync(ItemModel item)
     {
+        var existingItem = await dbContext.Items.FirstOrDefaultAsync(x => x.Id == item.Id);
+
+        if (existingItem is null)
+        {
+            return Error.NotFound(description: "Item not found!");
+        }
+
         var result = await dbContext.Items.AsNoTracking()
                                           .Where(x => x == item.ToEntity())
                                           .ExecuteDeleteAsync();
@@ -28,6 +36,7 @@ public class ItemService(AppDbContext dbContext) : IItemService
     public async Task<ErrorOr<ItemModel>> GetByIdAsync(int id)
     {
         var item = await dbContext.Items.FirstOrDefaultAsync(x => x.Id == id);
+
         if(item is null)
         {
             return Error.NotFound(description: "Item not found!");
@@ -37,6 +46,13 @@ public class ItemService(AppDbContext dbContext) : IItemService
 
     public async Task<ErrorOr<Success>> UpdateAsync(ItemModel item)
     {
+        var existingItem = await dbContext.Items.FirstOrDefaultAsync(x => x.Id == item.Id);
+
+        if (existingItem is null)
+        {
+            return Error.NotFound(description: "Item not found!");
+        }
+
         var result = await dbContext.Items.AsNoTracking()
                                           .Where(x => x.Id == item.Id)
                                           .ExecuteUpdateAsync(x => x.SetProperty(p => p.Name, item.Name)
